refactor(frontend): common component for refresh button
